+++ lib/commands/l10n/extract_command.dart
@@ -1,8 +1,7 @@
-import 'dart:io';
+
 
 import 'package:alex/alex.dart';
 import 'package:alex/src/exception/run_exception.dart';
-import 'package:path/path.dart' as path;
 
 import 'src/l10n_command_base.dart';
 
@@ -10,15 +9,6 @@ import 'src/l10n_command_base.dart';
 class ExtractCommand extends L10nCommandBase {
   ExtractCommand()
       : super('extract', 'Extract strings from Dart code to arb file.');
-
-  Future<File> getMainArb(L10nConfig l10nConfig) async {
-    final mainFile = _arb(L10nUtils.getArbMessagesFile(l10nConfig));
-    final localeFile = _arb(L10nUtils.getBaseArbFile(l10nConfig));
-
-    if (await localeFile.exists()) await localeFile.delete();
-    final res = await mainFile.copy(localeFile.path);
-    return res;
-  }
 
   @override
   Future<int> doRun() async {
@@ -40,12 +30,9 @@ class ExtractCommand extends L10nCommandBase {
       return errorBy(e);
     }
     
-    final mainFile = await getMainArb(l10nConfig);
+    final mainFile = await L10nUtils.getMainArb(l10nConfig);
     return success(
         message: 'Strings extracted to ARB file. '
             'You can send $mainFile to the translators');
   }
-
-  File _arb(String fileName) =>
-      File(path.join(config.l10n.outputDir, fileName));
 }