+++ lib/src/l10n/l10n_utils.dart
@@ -1,5 +1,9 @@
+
+import 'dart:io';
+
 import 'package:alex/alex.dart';
 import 'package:path/path.dart' as path;
+
 
 /// Localization utils.
 class L10nUtils {
@@ -42,6 +46,18 @@ class L10nUtils {
         : _getBaseNameByArb(config);
     return path.setExtension(fileName, '.xml');
   }
+
+  static  Future<File> getMainArb(L10nConfig l10nConfig) async {
+    final mainFile = _arb(L10nUtils.getArbMessagesFile(l10nConfig), l10nConfig);
+    final localeFile = _arb(L10nUtils.getBaseArbFile(l10nConfig), l10nConfig);
+
+    if (await localeFile.exists()) await localeFile.delete();
+    final res = await mainFile.copy(localeFile.path);
+    return res;
+  }
+
+ static File _arb(String fileName, L10nConfig config) =>
+      File(path.join(config.outputDir, fileName));
 
   static String _getBaseNameByArb(L10nConfig config) {
     final parts = getArbFileParts(config);