+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.5.0 - 2022-11-29
+
+* [L10n] Added `check_translate` command to check for translations for all locale strings.
+
+* [Release] Automatically check translations for locale (en by default) before run release.
+
 ## 1.4.1 - 2022-10-26
 
 * [L10n] import and export only difference strings of xml file.